requirements.txt
click==8.0.1
PyYAML>=6.0